+++ odoo/local-src/qoqa_offer/__init__.py
@@ -6,6 +6,6 @@ from . import stock
 from . import qoqa_offer
 from . import sale_order
 from . import product
-from . import procurement
 from . import orderpoint
 from . import res_users
+from . import wizard